# docs/src/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Message Vault HTTP API",
    "description": "HTTP API for a local Message Vault. Bearer session tokens come from login. API tokens come from Settings → Account.",
    "license": {
      "name": "Fair Core License 1.0 (ALv2 future)",
      "url": "https://github.com/bitrealm-io/message-vault/blob/main/LICENSE.md"
    },
    "version": "0.8.3"
  },
  "paths": {
...
    "/v1/account/api-tokens": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "List the account's named API tokens with their permissions and masked secrets.",
        "operationId": "list_api_tokens_handler",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListApiTokensResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Create a named API token. Returns the plaintext secret once, at creation;\nit is never returned again.",
        "operationId": "create_api_token_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/account/api-tokens/{id}": {
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Delete one named API token. Requests using it start failing on the next call.",
        "operationId": "delete_api_token_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "API token id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteApiTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Account"
        ],
        "summary": "Rename one named API token. The label is trimmed before storing.",
        "operationId": "rename_api_token_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "API token id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameApiTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenameApiTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiTokenItem": {
        "type": "object",
        "description": "One named API token as shown in Settings: label, permissions, and masked secret.",
        "required": [
          "id",
          "label",
          "can_import",
          "can_export",
          "can_delete",
          "token_hint",
          "created_at",
          "disabled"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean",
            "description": "May destroy message data."
          },
          "can_export": {
            "type": "boolean",
            "description": "May call the export endpoints."
          },
          "can_import": {
            "type": "boolean",
            "description": "May call the import endpoints."
          },
          "created_at": {
            "type": "string",
            "description": "Creation time as a Unix-seconds string."
          },
          "disabled": {
            "type": "boolean",
            "description": "True when the token is disabled and rejects requests."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Unix-seconds expiry; absent means no expiry."
          },
          "id": {
            "type": "string",
            "description": "Token id (the secret itself is stored hashed)."
          },
          "label": {
            "type": "string",
            "description": "User-chosen label shown in Settings."
          },
          "last_accessed_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Unix-seconds string of last use; absent when never used."
          },
          "token_hint": {
            "type": "string",
            "description": "Masked secret for Settings (e.g. `mv-api-Sd..mE`)."
          }
        }
      },
...
      "CreateApiTokenRequest": {
        "type": "object",
        "description": "Body for creating a token: label, permissions, optional expiry.",
        "required": [
          "label"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean",
            "description": "May destroy message data. Default false — asked for, never inherited."
          },
          "can_export": {
            "type": "boolean",
            "description": "May call the export endpoints. Default true."
          },
          "can_import": {
            "type": "boolean",
            "description": "May call the import endpoints. Default true."
          },
          "expires_in_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Days until expiry. Omit for the default (365 days). Pass `0` for no expiry.",
            "minimum": 0
          },
          "label": {
            "type": "string",
            "description": "User-chosen label shown in Settings."
          }
        }
      },
      "CreateApiTokenResponse": {
        "type": "object",
        "description": "The created token, including its plaintext secret (returned once).",
        "required": [
          "id",
          "label",
          "can_import",
          "can_export",
          "can_delete",
          "created_at",
          "token",
          "token_hint"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean",
            "description": "May destroy message data."
          },
          "can_export": {
            "type": "boolean",
            "description": "May call the export endpoints."
          },
          "can_import": {
            "type": "boolean",
            "description": "May call the import endpoints."
          },
          "created_at": {
            "type": "string",
            "description": "Creation time as a Unix-seconds string."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Unix-seconds expiry; absent means no expiry."
          },
          "id": {
            "type": "string",
            "description": "Token id."
          },
          "label": {
            "type": "string",
            "description": "User-chosen label."
          },
          "token": {
            "type": "string",
            "description": "Plaintext secret — returned once at creation."
          },
          "token_hint": {
            "type": "string",
            "description": "Masked form for the Settings list (also persisted)."
          }
        }
      },
...
      "DeleteApiTokenResponse": {
        "type": "object",
        "description": "Deletion acknowledgement.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Always true when a response is returned."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every failure: one sentence, with the HTTP status carrying the meaning.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable description of the failure."
          }
        }
      },
...
      "ListApiTokensResponse": {
        "type": "object",
        "description": "The account's named API tokens.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiTokenItem"
            },
            "description": "The account's tokens."
          }
        }
      },
...
      "RenameApiTokenRequest": {
        "type": "object",
        "description": "Body for renaming a token.",
        "required": [
          "label"
        ],
        "properties": {
          "label": {
            "type": "string",
            "description": "Replacement label."
          }
        }
      },
      "RenameApiTokenResponse": {
        "type": "object",
        "description": "The renamed token's id and stored label.",
        "required": [
          "ok",
          "id",
          "label"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Token id that was renamed."
          },
          "label": {
            "type": "string",
            "description": "Stored label after the rename."
          },
          "ok": {
            "type": "boolean",
            "description": "Always true when a response is returned."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}